Refactoring packaging
