+++ src/main/java/tf/gpx/edit/elevation/SRTMDataViewer_fxyz3d.java
@@ -67,9 +67,9 @@ import org.fxyz3d.scene.paint.Palette;
 import org.fxyz3d.shapes.composites.PolyLine3D;
 import org.fxyz3d.utils.CameraTransformer;
 import tf.gpx.edit.algorithms.EarthGeometry;
-import tf.gpx.edit.charts.Axis;
-import tf.gpx.edit.charts.Fxyz3dHelper;
-import tf.gpx.edit.charts.SurfacePlotMesh_Fast;
+import tf.gpx.edit.elevation.charts.Axis;
+import tf.gpx.edit.elevation.charts.Fxyz3dHelper;
+import tf.gpx.edit.elevation.charts.SurfacePlotMesh_Fast;
 import tf.gpx.edit.helper.GPXEditorPreferences;
 import tf.gpx.edit.items.Bounds3D;
 import tf.gpx.edit.items.GPXLineItem;
