+++ src/main/java/tf/gpx/edit/elevation/charts/Axis.java
@@ -23,7 +23,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  *  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tf.gpx.edit.charts;
+package tf.gpx.edit.elevation.charts;
 
 import javafx.scene.paint.Color;
 import javafx.scene.paint.PhongMaterial;
