+++ src/main/java/tf/gpx/edit/panorama/PanoramaViewer_Chart.java
@@ -58,7 +58,7 @@ import jfxtras.styles.jmetro.JMetro;
 import jfxtras.styles.jmetro.Style;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.lang3.tuple.Triple;
-import tf.gpx.edit.charts.SmoothFilledAreaChart;
+import tf.gpx.edit.elevation.charts.SmoothFilledAreaChart;
 import tf.gpx.edit.helper.LatLonHelper;
 import tf.gpx.edit.helper.TimeZoneProvider;
 import tf.gpx.edit.leafletmap.IGeoCoordinate;
